requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.1
attrs==25.4.0
bitarray==3.8.0
cbor2==5.8.0
certifi==2026.1.4
charset-normalizer==3.4.4
ckzg==2.1.5
comm==0.2.3
contourpy==1.3.3
crytic-compile==0.3.11
cycler==0.12.1
cytoolz==1.1.0
debugpy==1.8.20
decorator==5.2.1
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
executing==2.2.1
fonttools==4.61.1
frozenlist==1.8.0
hexbytes==1.3.1
idna==3.11
ipykernel==7.1.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
parsimonious==0.10.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyunormalize==17.0.0
pyzmq==27.1.0
regex==2026.1.15
requests==2.32.5
rlp==4.1.0
setuptools==80.10.2
six==1.17.0
slither-analyzer==0.11.5
solc-select==1.2.0
stack-data==0.6.3
toolz==1.1.0
tornado==6.5.4
traitlets==5.14.3
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.5.3
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
